feat: refactor headunit codebase with new hooks and services for battery, throttle, and lights data management
